Fix C ZKP-NL overflow (n=64) and CLI encfile KDF mismatch — v1.9.18 (TODO #79)
79.A: ZKP_NL_MAX_N was 32, so carry[33][3] overflowed by 30 rows when n=64.
Promoted all ZKP-NL share/carry types to uint64_t, bumped ZKP_NL_MAX_N to 64,
and fixed mask UB (1u<<n undefined for n>=32) with UINT64_MAX / 1ULL<<n.
Updated callers in herradura_cli.c, Herradura_tests.c, and the suite .c file.

79.B: cmd_encfile and cmd_decfile used ba_rol_k (v1.7 formula) instead of
ba_rnl_kdf_seed (v1.8.0, adds _RNL_KDF_DC XOR); C-generated .hkx files were
unreadable by Go/Python and vice versa. Two-line fix plus stale comment update.

+++ CryptosuiteTests/Herradura_tests.c
@@ -3632,7 +3632,7 @@ static void bench_zkp_nl(void)
     double secs;
     static const int n      = 32;
     static const int rounds = 16;
-    uint32_t A, B, y;
+    uint64_t A, B, y;
     ZkpNlRound *proof;
     int i;
     printf("[37] ZKP-NL prove+verify throughput  (n=%d, rounds=%d)  [PQC-EXT]\n",
@@ -3770,7 +3770,7 @@ static void test_zkp_nl_correctness(void)
         struct timespec t0;
         clock_gettime(CLOCK_MONOTONIC, &t0);
         for (i = 0; i < N; i++) {
-            uint32_t A, B, y;
+            uint64_t A, B, y;
             ZkpNlRound *proof;
             zkp_nl_keygen(n, urnd_fp, &A, &B, &y);
             proof = zkp_nl_prove(A, B, y, n, zkp_nl_rounds,
